% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{as.matrix.pdfium_bitmap}
\alias{as.matrix.pdfium_bitmap}
\title{Convert a pdfium_bitmap to a hex-color matrix}
\usage{
\method{as.matrix}{pdfium_bitmap}(x, ...)
}
\arguments{
\item{x}{A \code{pdfium_bitmap} from \code{\link[=pdf_render_page]{pdf_render_page()}}.}

\item{...}{Ignored.}
}
\value{
A character matrix of \code{"#RRGGBBAA"} colors.
}
\description{
Alias for \code{as.raster(x)}, included for symmetry with R's other
raster classes.
}
