% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_class.R
\name{as_pdfium_annot_list}
\alias{as_pdfium_annot_list}
\title{Coerce input to a \code{pdfium_annot_list}}
\usage{
as_pdfium_annot_list(x)
}
\arguments{
\item{x}{Either a \code{pdfium_annot_list}, a list of \code{pdfium_annot}
handles, or a tibble with a \code{handle} list-column.}
}
\value{
A \code{pdfium_annot_list}.
}
\description{
Reverse companion to \code{\link[=as_tibble.pdfium_annot_list]{as_tibble.pdfium_annot_list()}}: takes
either an existing list of \code{pdfium_annot} handles or a tibble
produced by \code{as_tibble()} and returns a \code{pdfium_annot_list}.
}
