% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{as_pdfium_bookmark_list}
\alias{as_pdfium_bookmark_list}
\title{Coerce input to a \code{pdfium_bookmark_list}}
\usage{
as_pdfium_bookmark_list(x)
}
\arguments{
\item{x}{Either a \code{pdfium_bookmark_list}, a list of
\code{pdfium_bookmark} handles, or a tibble with a \code{handle}
list-column.}
}
\value{
A \code{pdfium_bookmark_list}.
}
\description{
Reverse companion to \code{\link[=as_tibble.pdfium_bookmark_list]{as_tibble.pdfium_bookmark_list()}}.
}
