% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{as_pdfium_obj_list}
\alias{as_pdfium_obj_list}
\title{Coerce input to a \code{pdfium_obj_list}}
\usage{
as_pdfium_obj_list(x)
}
\arguments{
\item{x}{Either a \code{pdfium_obj_list}, a list of \code{pdfium_obj}
handles, or a tibble with a \code{handle} list-column.}
}
\value{
A \code{pdfium_obj_list}.
}
\description{
Reverse companion to \code{\link[=as_tibble.pdfium_obj_list]{as_tibble.pdfium_obj_list()}}.
}
