% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{as_pdfium_signature_list}
\alias{as_pdfium_signature_list}
\title{Coerce input to a \code{pdfium_signature_list}}
\usage{
as_pdfium_signature_list(x)
}
\arguments{
\item{x}{Either a \code{pdfium_signature_list}, a list of
\code{pdfium_signature} handles, or a tibble with a \code{handle}
list-column.}
}
\value{
A \code{pdfium_signature_list}.
}
\description{
Reverse companion to \code{\link[=as_tibble.pdfium_signature_list]{as_tibble.pdfium_signature_list()}}.
}
