% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_class.R
\name{pdf_annot_border_width}
\alias{pdf_annot_border_width}
\title{Annotation border width}
\usage{
pdf_annot_border_width(annot)
}
\arguments{
\item{annot}{A \code{pdfium_annot} handle from \code{\link[=pdf_annotations]{pdf_annotations()}}.}
}
\value{
Numeric scalar; \code{NA} if no border.
}
\description{
Returns the stroke border width from \verb{/Border} or \verb{/BS}.
}
