% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_class.R
\name{pdf_annot_subject}
\alias{pdf_annot_subject}
\title{Annotation \verb{/Subj} subject text}
\usage{
pdf_annot_subject(annot)
}
\arguments{
\item{annot}{A \code{pdfium_annot} handle from \code{\link[=pdf_annotations]{pdf_annotations()}}.}
}
\value{
Character scalar.
}
\description{
Returns the annotation's \verb{/Subj} subject string.
}
