% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_class.R
\name{pdf_annot_title}
\alias{pdf_annot_title}
\title{Annotation \verb{/T} title (author) text}
\usage{
pdf_annot_title(annot)
}
\arguments{
\item{annot}{A \code{pdfium_annot} handle from \code{\link[=pdf_annotations]{pdf_annotations()}}.}
}
\value{
Character scalar.
}
\description{
Returns the annotation's \verb{/T} title (commonly the author name).
}
