% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{pdf_bookmark_action_type}
\alias{pdf_bookmark_action_type}
\title{Bookmark action type}
\usage{
pdf_bookmark_action_type(bm)
}
\arguments{
\item{bm}{A \code{pdfium_bookmark} handle from \code{\link[=pdf_doc_bookmarks]{pdf_doc_bookmarks()}}.}
}
\value{
Character scalar.
}
\description{
Returns one of \code{"goto"}, \code{"remote_goto"}, \code{"uri"}, \code{"launch"},
\code{"embedded_goto"}, or \code{"unsupported"}. Wraps
\code{FPDFAction_GetType}.
}
