% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{pdf_bookmark_dest_view}
\alias{pdf_bookmark_dest_view}
\title{Bookmark destination view mode}
\usage{
pdf_bookmark_dest_view(bm)
}
\arguments{
\item{bm}{A \code{pdfium_bookmark} handle from \code{\link[=pdf_doc_bookmarks]{pdf_doc_bookmarks()}}.}
}
\value{
Character scalar.
}
\description{
Returns the destination view mode (one of \code{"xyz"}, \code{"fit"},
\code{"fith"}, \code{"fitv"}, \code{"fitr"}, \code{"fitb"}, \code{"fitbh"}, \code{"fitbv"},
\code{"unknown"}). Wraps \code{FPDFDest_GetView}.
}
