% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{pdf_obj_type}
\alias{pdf_obj_type}
\title{Report the type of a page object}
\usage{
pdf_obj_type(obj)
}
\arguments{
\item{obj}{A \code{pdfium_obj} from \code{\link[=pdf_page_objects]{pdf_page_objects()}}.}
}
\value{
Character scalar: one of \code{"path"}, \code{"text"}, \code{"image"},
\code{"form"}, \code{"shading"}, or \code{"unknown"}.
}
\description{
Report the type of a page object
}
