% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_setters.R
\name{pdf_path_set_line_cap}
\alias{pdf_path_set_line_cap}
\title{Set the line cap style of a path stroke}
\usage{
pdf_path_set_line_cap(obj, cap)
}
\arguments{
\item{obj}{A \code{pdfium_obj} of type \code{"path"}. Parent doc must be
readwrite.}

\item{cap}{Character scalar; one of \code{"butt"}, \code{"round"},
\code{"projecting_square"}.}
}
\value{
Invisibly returns the parent \code{pdfium_doc}.
}
\description{
Wraps \code{FPDFPageObj_SetLineCap}. Allowed values: \code{"butt"},
\code{"round"}, \code{"projecting_square"}.
}
\seealso{
\code{\link[=pdf_path_line_cap]{pdf_path_line_cap()}}.
}
