% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_setters.R
\name{pdf_path_set_line_join}
\alias{pdf_path_set_line_join}
\title{Set the line join style of a path stroke}
\usage{
pdf_path_set_line_join(obj, join)
}
\arguments{
\item{obj}{A \code{pdfium_obj} of type \code{"path"}. Parent doc must be
readwrite.}

\item{join}{Character scalar; one of \code{"miter"}, \code{"round"},
\code{"bevel"}.}
}
\value{
Invisibly returns the parent \code{pdfium_doc}.
}
\description{
Wraps \code{FPDFPageObj_SetLineJoin}. Allowed values: \code{"miter"},
\code{"round"}, \code{"bevel"}.
}
\seealso{
\code{\link[=pdf_path_line_join]{pdf_path_line_join()}}.
}
